training/pyproject.toml
[project]
name = "chess-nn-training"
version = "0.1.0"
description = "Training pipeline for chess ONNX evaluation networks"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "python-chess>=1.999",
    "numpy>=1.24",
    "onnx>=1.14",
    "onnxruntime>=1.16",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "modal>=0.64",
    "stockfish>=3.28",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
